Build main content, put photo, fix layout
I fixed the layout of the header, mainly height, padding and margins so that hopefully it looks good on all screen resolutions. Switched from px to % to make it more dynamic. Hope it works.
I built the main content div and put stock photos for all the cats and "cats".