Finish quotes & call for action divs
Built the html tags and css for the quotes and call for action div.

Moved all <a> css tags together because they're basically the same. So now also the nav bar links are blue on hover but I think it's nicer and more intuitive actually.

Added link to roboto font - looks much nicer! Available here in 100 italics, 300, 500, 700 and 900.

Only footer remaining now.